ProyectoBlog/requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
fontawesomefree==6.2.0
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
django-ckeditor==6.5.1
django-js-asset==2.0.0